Added (mix:), a macro for mixing two colours in different proportions using LCH.
This is heavily based off of CSS's color-mix(), but it uses a similar argument order to (gradient:). As for naming: I wanted to use (blend:) but came to realise that "blend" now only refers to blend modes, so I switched to (mix:). I was also contemplating having this explicitly be named (lch-mix:) but decided I didn't want to include the (hsl-mix:) that would be implied by such naming.

By the way, in Harlowe 4, I might switch the LCH-using macros to use the newer OKLCH instead.

Also:
 * Fixed a bug where (lch:) and all lch-related colour conversions were slightly wrong, due to the conversion algorithm assuming linear-light sRGB instead of gamut sRGB.
 * Fixed a bug where colours created via (lch:) could sometimes have r, g or b datavalues higher than 255 or lower than 0.
 * Fixed a bug where (background:) didn't work when given a Harlowe colour with fractional r, g, or b datavalues.
 * Now, instead of the docs removing the animated (text-style:) examples when Chrome is detected, they no longer animate in any browser unless the mouse is over their containing table.
 * Now, (bg:) is the main name for (background:), and the documentation refers to it mainly by that.