Added (p-not-before:).
This wasn't added previously because I was still thinking about how it would actually read, how it would be described in the documentation, and whether it would make sense alongside (p-not:) and so forth. I've convinced myself now that it's fine, though. After this, the only patterns really missing are (p-start:) and (p-end:), which weren't needed until the arrival of (str-replaced:).

Also added some tests for the previous commit's changes.

Closes #254.