Improved the behaviour of enchanting a hook's lines when those lines contain macros or links.
Now, instead of simply wrapping all of the text nodes in a line with a <tw-pseudo-hook>, which excluded <tw-expression>s that weren't rendered yet, it now wraps every element in a line, except where that element contains a <br> (whereupon only some of its contents are wrapped).

Note that enchanting lines still has ongoing issues involving inline elements that contain <br>s, which I want to fix backwards-incompatibly in 4.0.

Closes #249.