Worked out a new compromise for (line-style:) and line selection that can handle multi-line hooks more intuitively.
The big problem I'd been wrestling with is getting (line-style:) to "make sense" when wrapping a line that A) begins a hook halfway through, and B) where said hook spans multiple lines. Since it's not possible to jQuery wrap only the text nodes in the first half of the hook along with the text nodes in the same line outside of the hook without displacing some of the text nodes, I'd been forced to think of various solutions. In the end, I decided to compromise like so: the halves of a multi-line hook are now always treated as (and wrapped as) a separate line from the rest of the line. Even though that makes (text-indent:) not work quite right, other changers like (text-colour:) work correctly with this. It's not ideal, but I can't afford to keep fretting over this.

Also:

 * Fixed up more documentation lines.
 * 2.4 toolbar buttons now have more descriptive labels that TwineJS should use as tooltips.
 * Added an alternative error message for unmatched text tokens appearing in macro calls, in cases where it seems like a position name (like '1st') that was incorrectly used.
 * Code inside HTML comments is no longer syntax-highlighted.