Custom commands can now be serialised with (source:) and (v6m-source:), and are represented correctly in debug replays.
This representation is not guaranteed to be accurate throughout the story, for obvious reasons which are now laid out in the (source:) macro's article.

This also contains some refactoring needed to allow custom commands to be saved in save files, thus allowing every Harlowe value to be serialisable. More work on this will be in the next commit.